// src/types/components.types.ts
export interface HeaderMedia {
  hero_image: string;
  logo_image: string;
  whatsapp: string;
  instagram: string;
  shopee: string;
}

export interface FooterMedia {
  footer_image: string;
  logo_image: string;
  background_image: string;
  address: string;
  email: string;
  phone: number;
  maps: string;
  whatsapp: string;
  shopee: string;
  instagram: string;
}

export interface SocmedMedia {
  className?: string;
  whatsapp: string;
  instagram: string;
  shopee: string;
}